move the logic of generating the commit.id.describe-short value into the (abstract) GitDataProvider
